+++ internal/auth/auth.go
@@ -5,10 +5,18 @@ import (
 	"crypto/md5"
 	"crypto/rand"
 	"encoding/hex"
+	"errors"
 	"net/http"
 )
 
-const size = 4
+const (
+	size              = 4
+	userCookie        = "user_id"
+	noUserIDError     = "unable to verify user, no user_id provided"
+	verifyError       = "unable to verify user via sign"
+	IDError           = "unable to create user_id"
+	signCreationError = "unable to sign cookie"
+)
 
 type BasicAuth struct {
 	SecretKey string
@@ -20,32 +28,41 @@ func NewBasicAuth(secretKey string) *BasicAuth {
 	}
 }
 
-func (b *BasicAuth) CreateRandomID() string {
+func (b *BasicAuth) CreateRandomID() (string, error) {
 	sl := make([]byte, size)
-	rand.Read(sl)
-	return hex.EncodeToString(sl)
+	_, err := rand.Read(sl)
+	if err != nil {
+		return "", err
+	}
+	return hex.EncodeToString(sl), nil
 }
 
-func (b *BasicAuth) CreateNewSign() string {
+func (b *BasicAuth) CreateNewSign() (string, error) {
 	hash1 := hmac.New(md5.New, []byte(b.SecretKey))
-	hash1.Write([]byte(b.SecretKey))
-	sign := hash1.Sum(nil)
-	return hex.EncodeToString(sign)
+	_, err := hash1.Write([]byte(b.SecretKey))
+	if err != nil {
+		return "", err
+	}
+	return hex.EncodeToString(hash1.Sum(nil)), nil
 }
 
-func (b *BasicAuth) Authenticate(h http.HandlerFunc) http.HandlerFunc {
+func (b *BasicAuth) AuthentificateUser(h http.HandlerFunc) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		userID, err := r.Cookie("user_id")
-		if err != nil {
-			http.Error(w, "unable to verify user, no user_id provided", http.StatusUnauthorized)
+		userID, err := r.Cookie(userCookie)
+		if errors.Is(err, http.ErrNoCookie) {
+			http.Error(w, noUserIDError, http.StatusUnauthorized)
 			return
 		} else {
 			hash2 := hmac.New(md5.New, []byte(b.SecretKey))
-			hash2.Write([]byte(b.SecretKey))
+			_, err = hash2.Write([]byte(b.SecretKey))
+			if err != nil {
+				http.Error(w, verifyError, http.StatusUnauthorized)
+				return
+			}
 			sign := hash2.Sum(nil)
-			cook, _ := hex.DecodeString(userID.Value)
-			if !hmac.Equal(sign, cook[size:]) {
-				http.Error(w, "unable to verify user via sign", http.StatusUnauthorized)
+			cookie, err := hex.DecodeString(userID.Value)
+			if err != nil || !hmac.Equal(sign, cookie[size:]) {
+				http.Error(w, verifyError, http.StatusUnauthorized)
 				return
 			}
 		}
@@ -53,24 +70,50 @@ func (b *BasicAuth) Authenticate(h http.HandlerFunc) http.HandlerFunc {
 	}
 }
 
-func (b *BasicAuth) CheckCookies(h http.HandlerFunc) http.HandlerFunc {
+func (b *BasicAuth) CheckUserCookies(h http.HandlerFunc) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		userID, err := r.Cookie("user_id")
-		if err != nil {
-			cook := b.CreateRandomID() + b.CreateNewSign()
+		userID, err := r.Cookie(userCookie)
+		if errors.Is(err, http.ErrNoCookie) {
+			id, err := b.CreateRandomID()
+			if err != nil {
+				http.Error(w, IDError, http.StatusUnauthorized)
+				return
+			}
+
+			sign, err := b.CreateNewSign()
+			if err != nil {
+				http.Error(w, signCreationError, http.StatusUnauthorized)
+				return
+			}
+
+			cookieVal := id + sign
 			http.SetCookie(w, &http.Cookie{
-				Name:  "user_id",
-				Value: cook,
+				Name:  userCookie,
+				Value: cookieVal,
 			})
 		} else {
 			hash2 := hmac.New(md5.New, []byte(b.SecretKey))
 			hash2.Write([]byte(b.SecretKey))
 			sign := hash2.Sum(nil)
-			cook, _ := hex.DecodeString(userID.Value)
-			if !hmac.Equal(cook[size:], sign) {
+			cook, err := hex.DecodeString(userID.Value)
+
+			if err != nil || !hmac.Equal(cook[size:], sign) {
+				id, err := b.CreateRandomID()
+				if err != nil {
+					http.Error(w, IDError, http.StatusUnauthorized)
+					return
+				}
+
+				newSign, err := b.CreateNewSign()
+				if err != nil {
+					http.Error(w, signCreationError, http.StatusUnauthorized)
+					return
+				}
+
+				cookieVal := id + newSign
 				http.SetCookie(w, &http.Cookie{
-					Name:  "user_id",
-					Value: b.CreateRandomID() + b.CreateNewSign(),
+					Name:  userCookie,
+					Value: cookieVal,
 				})
 			}
 		}
