+++ internal/server/server.go
@@ -19,11 +19,10 @@ import (
 )
 
 type SimpleServer server.SimpleServer
-type Server server.Server
 
 func NewSimpleServer(conf config.Config, storage *os.File) *SimpleServer {
 	return &SimpleServer{Host: conf.Host, BaseURL: conf.BaseURL, Storage: storage, URLmap: map[string]string{},
-		UserURLS: map[string][]string{}, ID: 1, Config: conf}
+		UserURLS: map[string][]string{}, ID: 1, Config: conf, Compressor: &compressor.Compressor{}}
 }
 
 func (s *SimpleServer) CheckFileStorage() error {
@@ -44,11 +43,11 @@ func (s *SimpleServer) CheckFileStorage() error {
 
 func (s *SimpleServer) CheckDBStorage() error {
 	c := dbconnector.NewConnector(s.Config.DatabaseDSN)
-	err := c.Connect(func(db *sql.DB, args ...interface{}) error {
+	err := c.ConnectToDB(func(db *sql.DB, args ...interface{}) error {
 		return c.CreateTable(db)
 	})
 	if err == nil {
-		err := c.Connect(func(db *sql.DB, args ...interface{}) error {
+		err := c.ConnectToDB(func(db *sql.DB, args ...interface{}) error {
 			return c.ReadFromDB(db)
 		})
 		if err == nil {
@@ -60,17 +59,37 @@ func (s *SimpleServer) CheckDBStorage() error {
 	return err
 }
 
+func (s *SimpleServer) RouteHandlers(router *chi.Mux, hh handlers.HandlerHelper) {
+	ba := auth.NewBasicAuth(s.Config.SecretKey)
+
+	router.HandleFunc("/", ba.CheckUserCookies(s.Compressor.Decompress(
+		hh.ZapLogger.LogRequest(hh.HandlePostURL()))))
+
+	router.HandleFunc("/api/shorten", ba.CheckUserCookies(s.Compressor.Decompress(
+		hh.ZapLogger.LogRequest(hh.HandlePostAPIShorten()))))
+
+	router.HandleFunc("/api/shorten/batch", ba.CheckUserCookies(s.Compressor.Decompress(
+		hh.ZapLogger.LogRequest(hh.HandlePostAPIShortenBatch()))))
+
+	router.HandleFunc("/ping", hh.HandlePing())
+
+	router.Get("/api/user/urls", ba.AuthentificateUser(hh.ZapLogger.LogRequest(hh.HandleGetAPIUserURLs())))
+	router.Delete("/api/user/urls", ba.AuthentificateUser(hh.ZapLogger.LogRequest(hh.HandleDeleteAPIUserURLs())))
+}
+
 func (s *SimpleServer) RunServer() {
-	router := chi.NewRouter()
 	var hh handlers.HandlerHelper
-	hh.Config = s.Config
-	hh.Connector = dbconnector.NewConnector(hh.Config.DatabaseDSN)
 	var mylogger mylogger.Mylogger
+	var baseURL string
+	router := chi.NewRouter()
 
 	err := mylogger.Initialize("INFO")
 	if err != nil {
 		log.Fatal(err)
 	}
+	hh.Config = s.Config
+	hh.Connector = dbconnector.NewConnector(hh.Config.DatabaseDSN)
+
 	serv := server.SimpleServer{
 		Host:    s.Host,
 		BaseURL: s.BaseURL,
@@ -81,34 +100,22 @@ func (s *SimpleServer) RunServer() {
 	}
 	serv.URLsToUpdate = make(chan string, 100)
 	defer close(serv.URLsToUpdate)
+
 	hh.Server = &serv
-	hh.Z = mylogger
+	hh.ZapLogger = mylogger
 	hh.Connector.Z = &mylogger
 	hh.Router = router
 	hh.UserURLS = s.UserURLS
 
-	var baseURL string
-	ba := auth.NewBasicAuth(s.Config.SecretKey)
 	if s.BaseURL != "" {
 		baseURL = s.BaseURL + "/"
 	}
 	for k, v := range s.URLmap {
-		router.Get("/"+baseURL+k, compressor.Compress(mylogger.LogRequest(hh.HandleGetPostedURL("/"+k, v))))
+		router.Get("/"+baseURL+k, s.Compressor.Compress(mylogger.LogRequest(hh.HandleGetPostedURL("/"+k, v))))
 	}
-	router.HandleFunc("/", ba.CheckCookies(compressor.Decompress(
-		mylogger.LogRequest(hh.HandlePostURL()))))
 
-	router.HandleFunc("/api/shorten", ba.CheckCookies(compressor.Decompress(
-		mylogger.LogRequest(hh.HandlePostAPIShorten()))))
+	s.RouteHandlers(router, hh)
 
-	router.HandleFunc("/ping", hh.HandlePing())
-
-	router.HandleFunc("/api/shorten/batch", ba.CheckCookies(compressor.Decompress(
-		mylogger.LogRequest(hh.HandlePostAPIShortenBatch()))))
-
-	router.Get("/api/user/urls", ba.Authenticate(mylogger.LogRequest(hh.HandleGetAPIUserURLs())))
-
-	router.Delete("/api/user/urls", ba.Authenticate(mylogger.LogRequest(hh.HandleDeleteAPIUserURLs())))
 	err = http.ListenAndServe(s.Host, router)
 	if err != nil {
 		mylogger.LogError(err)
